Added more per host configuration tags
Added strip_path, strip_query and prefer_destination flags to each
redirect. They default to False and can be set by add_redirect.

The strip_path option omits all pathinfo coming from the client and
strip_query is the equivalent for query string. The prefer_destination
option does not overwrite the query parameters from the destination
with ones from the client.

Documented all of this and injecting our own stuff to the redirects.